+++ deep_hedging_v0/hedging/env_hedge_call.py
@@ -57,7 +57,8 @@ class HedgeCall(HedgeBase):
         self.reset_portfolio()
         # Reset asset positions
         self.shares_held = np.zeros(
-            (self.num_paths, self.num_assets, self.num_strikes, self.num_steps + 1)
+            (self.num_paths, self.num_assets, self.num_strikes, self.num_steps + 1),
+            dtype=np.float32,
         )
         # Initial state features
         self.state[:, -1, :] = self._create_observations()
